+++ src/components/laundry.js
@@ -21,6 +21,12 @@ const Div = styled.div`
     span {
       margin-left: 1rem;
     }
+
+    @media only screen and (max-width: 1200px) {
+      span {
+        display: none;
+      }
+    }
   }
 
   &.animate {