+++ src/components/weather.js
@@ -68,6 +68,12 @@ const Div = styled.div`
       font-size: 1rem;
     }
   }
+
+  @media only screen and (max-width: 1200px) {
+    .forecast {
+      display: none;
+    }
+  }
   
   .detail-header {
     display: flex;
@@ -148,28 +154,18 @@ const Weather = ({ pin }) => {
   if (!data || !('currently' in data) || !('daily' in data) || !('hourly' in data)) return ''
 
   return (
-    <Div>
+    <Div onClick={() => toggle(true)}>
       <div className={'headline'}>
         <Icon icon={data.currently.icon}/>
-        <h2>{Math.round(data.currently.temperature)}° C</h2>
-      </div>
-      <div className={'values'}>
-        <Label icon={data.currently.icon} />
-        <div><span>Gefühlt:</span> {Math.round(data.currently.apparentTemperature)}° C</div>
-        <div><span>Luftfeuchtigkeit:</span> {data.currently.humidity * 100}%</div>
-        <div><span>Wind:</span> {Math.round(data.currently.windSpeed)} km/h</div>
-        <div><span>Bewölkung:</span> {data.currently.cloudCover * 100} %</div>
-        <div><span>Regen:</span> {data.currently.precipProbability * 1000} %</div>
-        <div><span>Sonnenaufgang:</span> {DateTime.fromSeconds(data.daily.data[0].sunriseTime).toLocaleString(DateTime.TIME_24_SIMPLE)}</div>
-        <div><span>Sonnenuntergang:</span> {DateTime.fromSeconds(data.daily.data[0].sunsetTime).toLocaleString(DateTime.TIME_24_SIMPLE)}</div>
+        <h2>{Math.round(data.currently.temperature)}°</h2>
       </div>
       <div className={'forecast'}>
         {[3,6,9,12].map((i, index) => (
           <Forecast key={index} data={data.hourly.data[i]} />
         ))}
       </div>
-      <Overlay visible={showWeather}>
-        <div className={'full-weather'}>
+      <Overlay visible={showWeather} onClick={() => toggle(false)}>
+        <div className={'full-weather'} onClick={() => toggle(false)}>
           <h2>Wetter Vorhersage</h2>
           <div className={'detail-header'}>
             <div>