+++ src/utils/use-calendar-data.js
@@ -84,16 +84,12 @@ const loadAll = (startDate, data, setData, toggleLoading) => {
   }
 }
 
-const useCalendarData = (startDate, setStartDate) => {
+const useCalendarData = (startDate) => {
   const [ data, setData ] = React.useState([])
   const timeout = useTimeout(60000, 'Calendar')
 
   React.useEffect(() => {
-    // If today is visible but not the first column, start week with today
-    const today = DateTime.now()
-    if (!startDate.hasSame(today, 'day') && startDate < today && today <= startDate.plus({ days: 6})) {
-      setStartDate(today)
-    } else {
+    if (startDate !== undefined) {
       loadAll(startDate, data, setData, () => { return })
     }
   // eslint-disable-next-line